backend/requirements.txt
requests==2.26.0
Django==3.2.25
djangorestframework==3.12.4
django-filter==23.5
python-dotenv==0.20.0
psycopg2-binary==2.9.3
djoser==2.2.2
gunicorn==20.1.0
Pillow==9.3.0
django-colorfield==0.11.0